requirements.txt
torch
numpy
pyyaml

